Maintainability: Unified requirements.txt file
