+++ .travis.yml
@@ -17,9 +17,8 @@ jobs:
       python: 3.6
 
 before_install:
-  - if [ "$TRAVIS_OS_NAME" = "osx" ]; then pip3 install -r requirements/requirements-osx.txt; fi
+  - pip3 -r requirements/requirements.txt
   - if [ "$TRAVIS_OS_NAME" = "linux" ]; then sudo apt-get install python3-pip pyqt5-dev-tools libomp-dev build-essential; fi
-  - if [ "$TRAVIS_OS_NAME" = "linux" ]; then pip3 install -r requirements/requirements-linux.txt; fi
 install:
   - pyrcc5 -o libs/resources.py resources.qrc
   - python3 setup.py build_ext --inplace
