--- requirements/requirements-osx.txt
@@ -1,9 +1,0 @@
-pyqt5>=5.12
-lxml>=4.2.4
-defusedxml>=0.6.0
-Cython==0.29.6
-opencv-contrib-python-headless==4.0.0.21
-tensorflow>=2.0
-numpy==1.16.2
-traces==0.4.2
-requests[darkmode]
